backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
celery
redis
alembic
pytest
httpx
